feat(api): story 33.14 - phpstan-strict-rules, 470 findings to zero

+++ api/src/Shared/Infrastructure/Http/DockerSocketClient.php
@@ -213,7 +213,7 @@ final readonly class DockerSocketClient
 
         $response = curl_exec($ch);
         $error = curl_error($ch);
-        $httpCode = (int) curl_getinfo($ch, \CURLINFO_HTTP_CODE);
+        $httpCode = curl_getinfo($ch, \CURLINFO_HTTP_CODE);
         curl_close($ch);
 
         if (!is_string($response)) {
